server/requirements.txt
Django==1.11.15
django-cors-headers==2.4.0
djangorestframework==3.8.2
djangorestframework-jsonapi==2.6.0
inflection==0.3.1
Markdown==3.0.1
psycopg2-binary==2.7.5
pytz==2018.5
six==1.11.0
